requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.3.0
requests==2.26.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
